Refactored LanguageDropdown
